ML/Deploy/requirements.txt
Flask==2.2.2
numpy==1.23.5
pandas==1.5.3
Pillow==9.5.0
Requests==2.31.0
scikit_learn==1.2.1
tensorflow==2.12.0
gunicorn==20.1.0
